Smarter Script Added some smarts to the script to return exit codes when the test tool doesn't work as expected. Also checks the library name for dynlib or so and chooses soft link names as appropriate. Removed the worktree use as it isn't supported on the test server. Added include signal.h to the test tool.

+++ abi-test.sh
@@ -9,22 +9,12 @@ _certs=(
     certs/server-cert.pem
     certs/server-key.pem
 )
-_oln="libwolfssl.3.dylib"
-_ln="libwolfssl.dylib"
-
-
-function runTest {
-    ./server -d &
-    _pid="$!"
-    ./a.out
-    kill -9 "$_pid"
-}
 
 
 if test ! -d wolfssl
 then
     echo "Fetching wolfssl from GitHub"
-    git clone git@github.com:wolfSSL/wolfssl.git
+    git clone https://github.com/wolfSSL/wolfssl.git
 fi
 
 echo "Cleanup from previous run"
@@ -37,39 +27,65 @@ git checkout master
 git fetch origin
 git reset --hard origin/master
 
-./autogen.sh
-./configure --disable-dependency-tracking --disable-shared
-make examples/server/server
+./autogen.sh >/dev/null 2>&1
+./configure --disable-dependency-tracking --disable-shared >/dev/null
+make examples/server/server >/dev/null
 cp examples/server/server "$_pwd"
 cp "${_certs[@]}" "$_pwd/certs"
 
-if test ! -d ABI_OLD
-then
-    git worktree add ABI_OLD "$_reftag"
-    pushd ABI_OLD
-    ./autogen.sh
-    ./configure --disable-dependency-tracking --disable-static --prefix="$_pwd/local"
-    popd
-fi
-
-cd ABI_OLD
-make install
+git checkout "$_reftag"
+./autogen.sh >/dev/null 2>&1
+./configure --disable-dependency-tracking --disable-static --prefix="$_pwd/local" >/dev/null
+make install >/dev/null
 popd
+
+echo "$(ls $_pwd/local/lib)"
+case "$(ls $_pwd/local/lib)" in
+*.dylib*)
+    _oln="libwolfssl.3.dylib"
+    _ln="libwolfssl.dylib"
+    ;;
+*.so*)
+    _oln="libwolfssl.3.so"
+    _ln="libwolfssl.so"
+    ;;
+esac
 
 gcc main.c -L./local/lib -I./local/include -lwolfssl
-runTest
+./server -d &
+if ! ./a.out
+then
+    echo "case 1: Expected success, failed. Fail."
+    exit 1
+fi
 
 rm -f "local/lib/$_oln"
-runTest
+if ./a.out
+then
+    echo "case 2: Expected failure, passed. Fail."
+    exit 1
+fi
 
 pushd wolfssl
-./configure --disable-dependency-tracking --disable-static --prefix="$_pwd/local"
-make install
+rm -f support/wolfssl.pc
+git checkout master
+./autogen.sh >/dev/null 2>&1
+./configure --disable-dependency-tracking --disable-static --prefix="$_pwd/local" >/dev/null
+make install >/dev/null
 popd
-runTest
+if ./a.out
+then
+    echo "case 3: Expected failure, passed. Fail."
+    exit 1
+fi
 
 pushd local/lib
 ln -sf "$_ln" "$_oln"
 popd
-runTest
+./server -d &
+if ! ./a.out
+then
+    echo "case 4: Expected success, failed. Fail."
+    exit 1
+fi
